Implementation-specific recognizer with gst_pocketsphinx
